The usbip tools use packed structs for network communication. Taking the address of a packed member of a struct can crash the program with SIGBUS on architectures with strict alignment requirements.
Also, recent versions of GCC detect this situation and emit a warning that
is fatal due to -Werror being used.

error: taking address of packed member of ‘struct
usbip_usb_device’ may result in an unaligned pointer value [-Werror=address-
of-packed-member]

Fix this by copying the data to an aligned location and operating there.